api/requirements.txt
# ─── Core ML & Embedding libs ─────────────────────────
sentence-transformers>=2.2.2,<3.0     # stable version with compatible deps
transformers>=4.21.0,<5.0              # compatible with sentence-transformers
huggingface-hub>=0.15.0,<1.0           # compatible hub version
scikit-learn>=1.3.0                    # for cosine similarity
numpy>=1.24.0                          # for array operations
torch>=2.1.0                           # required by sentence-transformers 2.7+

# ─── Database & environment ─────────────────────────────
python-dotenv>=1.0               # loads .env at startup

# ─── FastAPI server stack ─────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.7,<3.0
